+++ spec/dummy/app/javascript/application.js
@@ -1,2 +1,1 @@
-import './controllers/application.js'
 import "controllers"
